4.0. File [#252491]111

+++ src/main/java/ru/job4j/io/Dir.java
@@ -18,7 +18,7 @@ public class Dir {
         }
         System.out.println(String.format("size : %s", file.getTotalSpace()));
         for (File subfile : file.listFiles()) {
-            int sizeSubFile = volumeDirectory(subfile);
+            long sizeSubFile = volumeDirectory(subfile);
             System.out.println(String.valueOf(subfile.getAbsoluteFile()) + " " + sizeSubFile);
         }
     }
